Generalise `indices(of:)` to move it from `RandomAccessCollection` to `Collection`. Generalise `Collection` functions  `indices(where:)` and `forEach(slice:body:)` to work with any `Index` type. Generalise `Collection` function `average()` to work with any `BinaryInteger` elements.
