+++ Sources/SwifterSwift/SwiftStdlib/CollectionExtensions.swift
@@ -59,11 +59,6 @@ public extension Collection {
         return slices
     }
 
-}
-
-// MARK: - Methods (Int)
-public extension Collection where Index == Int {
-
     /// SwifterSwift: Get all indices where condition is met.
     ///
     ///     [1, 7, 1, 2, 4, 1, 8].indices(where: { $0 == 1 }) -> [0, 2, 5]
@@ -71,11 +66,8 @@ public extension Collection where Index == Int {
     /// - Parameter condition: condition to evaluate each element against.
     /// - Returns: all indices where the specified condition evaluates to true. (optional)
     func indices(where condition: (Element) throws -> Bool) rethrows -> [Index]? {
-        var indicies: [Index] = []
-        for (index, value) in lazy.enumerated() where try condition(value) {
-            indicies.append(index)
-        }
-        return indicies.isEmpty ? nil : indicies
+        let indices = try self.indices.filter { try condition(self[$0]) }
+        return indices.isEmpty ? nil :indices
     }
 
     /// SwifterSwift: Calls the given closure with an array of size of the parameter slice.
@@ -89,24 +81,42 @@ public extension Collection where Index == Int {
     func forEach(slice: Int, body: ([Element]) throws -> Void) rethrows {
         guard slice > 0, !isEmpty else { return }
 
-        var value: Int = 0
-        while value < count {
-            try body(Array(self[Swift.max(value, startIndex)..<Swift.min(value + slice, endIndex)]))
-            value += slice
+        var start = startIndex
+        while let end = index(start, offsetBy: slice, limitedBy: endIndex) {
+            try body(Array(self[start..<end]))
+            start = index(start, offsetBy: slice)
         }
     }
 
 }
 
-// MARK: - Methods (Integer)
-public extension Collection where Element == IntegerLiteralType, Index == Int {
+// MARK: - Methods (Equatable)
+public extension Collection where Element: Equatable {
+
+    /// SwifterSwift: All indices of specified item.
+    ///
+    ///        [1, 2, 2, 3, 4, 2, 5].indices(of 2) -> [1, 2, 5]
+    ///        [1.2, 2.3, 4.5, 3.4, 4.5].indices(of 2.3) -> [1]
+    ///        ["h", "e", "l", "l", "o"].indices(of "l") -> [2, 3]
+    ///
+    /// - Parameter item: item to check.
+    /// - Returns: an array with all indices of the given item.
+    func indices(of item: Element) -> [Index] {
+        return indices.filter { self[$0] == item }
+    }
+
+}
+
+// MARK: - Methods (BinaryInteger)
+public extension Collection where Element: BinaryInteger {
 
     /// SwifterSwift: Average of all elements in array.
     ///
     /// - Returns: the average of the array's elements.
     func average() -> Double {
         // http://stackoverflow.com/questions/28288148/making-my-function-calculate-average-of-array-swift
-        return isEmpty ? 0 : Double(reduce(0, +)) / Double(count)
+        guard !isEmpty else { return 0 }
+        return Double(reduce(.zero, +)) / Double(count)
     }
 
 }
@@ -121,7 +131,7 @@ public extension Collection where Element: FloatingPoint {
     /// - Returns: average of the array's elements.
     func average() -> Element {
         guard !isEmpty else { return 0 }
-        return reduce(0, {$0 + $1}) / Element(count)
+        return reduce(.zero, +) / Element(count)
     }
 
 }
